trader/requirements.txt
eth-account==0.13.7
python-dotenv==1.0.1
flask==3.0.0
web3==6.15.1
requests==2.31.0
pydantic==2.5.2
cryptography==41.0.8
gunicorn==21.2.0
